Fixed compiler warning.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@7867 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/graphics/lod_node.cpp
@@ -59,7 +59,7 @@ void LODNode::add(int level, scene::ISceneNode* node, bool reparent)
 {
     // samuncle suggested to put a slight randomisation in LOD
     // I'm not convinced (Auria) but he's the artist pro, so I listen ;P
-    level += ((rand()%1000)-500)/500.0f*(level*0.1f);
+    level += (int)(((rand()%1000)-500)/500.0f*(level*0.1f));
     
     // FIXME : this class assumes that 'm_detail' is sorted, this may not always be the case
     node->grab();
